+++ tests/components/feedreader/test_init.py
@@ -10,8 +10,7 @@ from unittest.mock import MagicMock, mock_open, patch
 
 import pytest
 
-from homeassistant.components import feedreader
-from homeassistant.components.feedreader import (
+from homeassistant.components.feedreader.const import (
     CONF_MAX_ENTRIES,
     CONF_URLS,
     DEFAULT_SCAN_INTERVAL,
@@ -26,11 +25,11 @@ import homeassistant.util.dt as dt_util
 from tests.common import async_capture_events, async_fire_time_changed, load_fixture
 
 URL = "http://some.rss.local/rss_feed.xml"
-VALID_CONFIG_1 = {feedreader.DOMAIN: {CONF_URLS: [URL]}}
-VALID_CONFIG_2 = {feedreader.DOMAIN: {CONF_URLS: [URL], CONF_SCAN_INTERVAL: 60}}
-VALID_CONFIG_3 = {feedreader.DOMAIN: {CONF_URLS: [URL], CONF_MAX_ENTRIES: 100}}
-VALID_CONFIG_4 = {feedreader.DOMAIN: {CONF_URLS: [URL], CONF_MAX_ENTRIES: 5}}
-VALID_CONFIG_5 = {feedreader.DOMAIN: {CONF_URLS: [URL], CONF_MAX_ENTRIES: 1}}
+VALID_CONFIG_1 = {DOMAIN: {CONF_URLS: [URL]}}
+VALID_CONFIG_2 = {DOMAIN: {CONF_URLS: [URL], CONF_SCAN_INTERVAL: 60}}
+VALID_CONFIG_3 = {DOMAIN: {CONF_URLS: [URL], CONF_MAX_ENTRIES: 100}}
+VALID_CONFIG_4 = {DOMAIN: {CONF_URLS: [URL], CONF_MAX_ENTRIES: 5}}
+VALID_CONFIG_5 = {DOMAIN: {CONF_URLS: [URL], CONF_MAX_ENTRIES: 1}}
 
 
 def load_fixture_bytes(src: str) -> bytes:
@@ -98,7 +97,7 @@ def fixture_storage(request: pytest.FixtureRequest) -> Generator[None, None, Non
 def fixture_legacy_storage_open() -> Generator[MagicMock, None, None]:
     """Mock builtins.open for feedreader storage."""
     with patch(
-        "homeassistant.components.feedreader.open",
+        "homeassistant.components.feedreader.coordinator.open",
         mock_open(),
         create=True,
     ) as open_mock:
@@ -111,16 +110,14 @@ def fixture_legacy_storage_load(
 ) -> Generator[MagicMock, None, None]:
     """Mock builtins.open for feedreader storage."""
     with patch(
-        "homeassistant.components.feedreader.pickle.load", return_value={}
+        "homeassistant.components.feedreader.coordinator.pickle.load", return_value={}
     ) as pickle_load:
         yield pickle_load
 
 
 async def test_setup_no_feeds(hass: HomeAssistant) -> None:
     """Test config with no urls."""
-    assert not await async_setup_component(
-        hass, feedreader.DOMAIN, {feedreader.DOMAIN: {CONF_URLS: []}}
-    )
+    assert not await async_setup_component(hass, DOMAIN, {DOMAIN: {CONF_URLS: []}})
 
 
 @pytest.mark.parametrize(
@@ -143,9 +140,9 @@ async def test_legacy_storage_error(
     legacy_storage_load.side_effect = load_error
 
     with patch(
-        "homeassistant.components.feedreader.async_track_time_interval"
+        "homeassistant.components.feedreader.coordinator.async_track_time_interval"
     ) as track_method:
-        assert await async_setup_component(hass, feedreader.DOMAIN, VALID_CONFIG_1)
+        assert await async_setup_component(hass, DOMAIN, VALID_CONFIG_1)
         await hass.async_block_till_done()
 
     track_method.assert_called_once_with(
@@ -164,10 +161,10 @@ async def test_storage_data_loading(
 ) -> None:
     """Test loading existing storage data."""
     storage_data: dict[str, str] = {URL: "2018-04-30T05:10:00+00:00"}
-    hass_storage[feedreader.DOMAIN] = {
+    hass_storage[DOMAIN] = {
         "version": 1,
         "minor_version": 1,
-        "key": feedreader.DOMAIN,
+        "key": DOMAIN,
         "data": storage_data,
     }
     legacy_storage_data = {
@@ -179,7 +176,7 @@ async def test_storage_data_loading(
         "feedparser.http.get",
         return_value=feed_one_event,
     ):
-        assert await async_setup_component(hass, feedreader.DOMAIN, VALID_CONFIG_2)
+        assert await async_setup_component(hass, DOMAIN, VALID_CONFIG_2)
 
         hass.bus.async_fire(EVENT_HOMEASSISTANT_START)
         await hass.async_block_till_done()
@@ -202,9 +199,9 @@ async def test_storage_data_writing(
             "feedparser.http.get",
             return_value=feed_one_event,
         ),
-        patch("homeassistant.components.feedreader.DELAY_SAVE", new=0),
+        patch("homeassistant.components.feedreader.coordinator.DELAY_SAVE", new=0),
     ):
-        assert await async_setup_component(hass, feedreader.DOMAIN, VALID_CONFIG_2)
+        assert await async_setup_component(hass, DOMAIN, VALID_CONFIG_2)
 
         hass.bus.async_fire(EVENT_HOMEASSISTANT_START)
         await hass.async_block_till_done()
@@ -213,16 +210,16 @@ async def test_storage_data_writing(
     assert len(events) == 1
 
     # storage data updated
-    assert hass_storage[feedreader.DOMAIN]["data"] == storage_data
+    assert hass_storage[DOMAIN]["data"] == storage_data
 
 
 @pytest.mark.parametrize("storage", ["legacy_storage", "json_storage"], indirect=True)
 async def test_setup_one_feed(hass: HomeAssistant, storage: None) -> None:
     """Test the general setup of this component."""
     with patch(
-        "homeassistant.components.feedreader.async_track_time_interval"
+        "homeassistant.components.feedreader.coordinator.async_track_time_interval"
     ) as track_method:
-        assert await async_setup_component(hass, feedreader.DOMAIN, VALID_CONFIG_1)
+        assert await async_setup_component(hass, DOMAIN, VALID_CONFIG_1)
         await hass.async_block_till_done()
 
     track_method.assert_called_once_with(
@@ -233,9 +230,9 @@ async def test_setup_one_feed(hass: HomeAssistant, storage: None) -> None:
 async def test_setup_scan_interval(hass: HomeAssistant) -> None:
     """Test the setup of this component with scan interval."""
     with patch(
-        "homeassistant.components.feedreader.async_track_time_interval"
+        "homeassistant.components.feedreader.coordinator.async_track_time_interval"
     ) as track_method:
-        assert await async_setup_component(hass, feedreader.DOMAIN, VALID_CONFIG_2)
+        assert await async_setup_component(hass, DOMAIN, VALID_CONFIG_2)
         await hass.async_block_till_done()
 
     track_method.assert_called_once_with(
@@ -245,7 +242,7 @@ async def test_setup_scan_interval(hass: HomeAssistant) -> None:
 
 async def test_setup_max_entries(hass: HomeAssistant) -> None:
     """Test the setup of this component with max entries."""
-    assert await async_setup_component(hass, feedreader.DOMAIN, VALID_CONFIG_3)
+    assert await async_setup_component(hass, DOMAIN, VALID_CONFIG_3)
     await hass.async_block_till_done()
 
 
@@ -255,7 +252,7 @@ async def test_feed(hass: HomeAssistant, events, feed_one_event) -> None:
         "feedparser.http.get",
         return_value=feed_one_event,
     ):
-        assert await async_setup_component(hass, feedreader.DOMAIN, VALID_CONFIG_2)
+        assert await async_setup_component(hass, DOMAIN, VALID_CONFIG_2)
 
         hass.bus.async_fire(EVENT_HOMEASSISTANT_START)
         await hass.async_block_till_done()
@@ -278,7 +275,7 @@ async def test_atom_feed(hass: HomeAssistant, events, feed_atom_event) -> None:
         "feedparser.http.get",
         return_value=feed_atom_event,
     ):
-        assert await async_setup_component(hass, feedreader.DOMAIN, VALID_CONFIG_5)
+        assert await async_setup_component(hass, DOMAIN, VALID_CONFIG_5)
 
         hass.bus.async_fire(EVENT_HOMEASSISTANT_START)
         await hass.async_block_till_done()
@@ -305,13 +302,13 @@ async def test_feed_identical_timestamps(
             return_value=feed_identically_timed_events,
         ),
         patch(
-            "homeassistant.components.feedreader.StoredData.get_timestamp",
+            "homeassistant.components.feedreader.coordinator.StoredData.get_timestamp",
             return_value=gmtime(
                 datetime.fromisoformat("1970-01-01T00:00:00.0+0000").timestamp()
             ),
         ),
     ):
-        assert await async_setup_component(hass, feedreader.DOMAIN, VALID_CONFIG_2)
+        assert await async_setup_component(hass, DOMAIN, VALID_CONFIG_2)
 
         hass.bus.async_fire(EVENT_HOMEASSISTANT_START)
         await hass.async_block_till_done()
@@ -366,7 +363,7 @@ async def test_feed_updates(
     ]
 
     with patch("feedparser.http.get", side_effect=side_effect):
-        assert await async_setup_component(hass, feedreader.DOMAIN, VALID_CONFIG_2)
+        assert await async_setup_component(hass, DOMAIN, VALID_CONFIG_2)
 
         hass.bus.async_fire(EVENT_HOMEASSISTANT_START)
         await hass.async_block_till_done()
@@ -393,7 +390,7 @@ async def test_feed_default_max_length(
 ) -> None:
     """Test long feed beyond the default 20 entry limit."""
     with patch("feedparser.http.get", return_value=feed_21_events):
-        assert await async_setup_component(hass, feedreader.DOMAIN, VALID_CONFIG_2)
+        assert await async_setup_component(hass, DOMAIN, VALID_CONFIG_2)
 
         hass.bus.async_fire(EVENT_HOMEASSISTANT_START)
         await hass.async_block_till_done()
@@ -404,7 +401,7 @@ async def test_feed_default_max_length(
 async def test_feed_max_length(hass: HomeAssistant, events, feed_21_events) -> None:
     """Test long feed beyond a configured 5 entry limit."""
     with patch("feedparser.http.get", return_value=feed_21_events):
-        assert await async_setup_component(hass, feedreader.DOMAIN, VALID_CONFIG_4)
+        assert await async_setup_component(hass, DOMAIN, VALID_CONFIG_4)
 
         hass.bus.async_fire(EVENT_HOMEASSISTANT_START)
         await hass.async_block_till_done()
@@ -417,7 +414,7 @@ async def test_feed_without_publication_date_and_title(
 ) -> None:
     """Test simple feed with entry without publication date and title."""
     with patch("feedparser.http.get", return_value=feed_three_events):
-        assert await async_setup_component(hass, feedreader.DOMAIN, VALID_CONFIG_2)
+        assert await async_setup_component(hass, DOMAIN, VALID_CONFIG_2)
 
         hass.bus.async_fire(EVENT_HOMEASSISTANT_START)
         await hass.async_block_till_done()
@@ -432,7 +429,7 @@ async def test_feed_with_unrecognized_publication_date(
     with patch(
         "feedparser.http.get", return_value=load_fixture_bytes("feedreader4.xml")
     ):
-        assert await async_setup_component(hass, feedreader.DOMAIN, VALID_CONFIG_2)
+        assert await async_setup_component(hass, DOMAIN, VALID_CONFIG_2)
 
         hass.bus.async_fire(EVENT_HOMEASSISTANT_START)
         await hass.async_block_till_done()
@@ -444,7 +441,7 @@ async def test_feed_invalid_data(hass: HomeAssistant, events) -> None:
     """Test feed with invalid data."""
     invalid_data = bytes("INVALID DATA", "utf-8")
     with patch("feedparser.http.get", return_value=invalid_data):
-        assert await async_setup_component(hass, feedreader.DOMAIN, VALID_CONFIG_2)
+        assert await async_setup_component(hass, DOMAIN, VALID_CONFIG_2)
 
         hass.bus.async_fire(EVENT_HOMEASSISTANT_START)
         await hass.async_block_till_done()
@@ -459,7 +456,7 @@ async def test_feed_parsing_failed(
     assert "Error fetching feed data" not in caplog.text
 
     with patch("feedparser.parse", return_value=None):
-        assert await async_setup_component(hass, feedreader.DOMAIN, VALID_CONFIG_2)
+        assert await async_setup_component(hass, DOMAIN, VALID_CONFIG_2)
 
         hass.bus.async_fire(EVENT_HOMEASSISTANT_START)
         await hass.async_block_till_done()
